refactor: 확장 함수 분리 및 네이밍 변경
