+++ app/src/main/java/woowacourse/movie/extensions/IntentExtension.kt
@@ -1,0 +1,12 @@
+package woowacourse.movie.extensions
+
+import android.content.Intent
+import android.os.Build
+import android.os.Parcelable
+
+inline fun <reified T : Parcelable> Intent.getParcelableCompat(key: String): T? =
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+        getParcelableExtra(key, T::class.java)
+    } else {
+        getParcelableExtra(key) as? T
+    }
